functions/requirements.txt
stream-python==5.1.0
flask==2.0.2
firebase-admin==5.0.3